% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{boxplot_AUC}
\alias{boxplot_AUC}
\title{Boxplot of AUC distribution across repetitions}
\usage{
boxplot_AUC(curve_data, auc_limit = 1)
}
\arguments{
\item{curve_data}{the ouput of the \code{\link{generate_curve_data}} function}
}
\value{
a ggplot object with the boxplot data
}
\description{
Boxplot of AUC distribution across repetitions
}
