% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_features.R
\name{boxplot_rank}
\alias{boxplot_rank}
\title{Plot the distribution of the rank accross methods}
\usage{
boxplot_rank(dt, nmax = 30)
}
\arguments{
\item{nmax}{max number of features to plot}

\item{data}{the output of the \code{\link{rank_features}} function}
}
\value{
boxplot. ggplot object
}
\description{
Plot the distribution of the rank accross methods
}
