% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_DAF_data.R
\name{build_DAF_data}
\alias{build_DAF_data}
\title{Load all data for DAF analysis}
\usage{
build_DAF_data(count_table, metadata, group_var = NULL,
  simulation = TRUE, ...)
}
\arguments{
\item{count_table}{a matrix of count values with features as rows and samples as columns.}

\item{metadata}{a data frame. Rows are samples and miscellaneous variables as columns.}

\item{group_var}{character. The name of \code{metadata} column defining the groups.}

\item{simulation}{logical. Should differences be generated ? Default is TRUE.}

\item{...}{additionnal parameter passed to the \code{\link{generateData}} function}
}
\value{
a list of 2 or 4 elements :
  \item{count_table}{ a matrix of count values.}
  \item{metadata}{ a data frame of at least one column countaining the group variable.}

  If the \code{simulation} parameter was TRUE, artificial differences were
  generated. It is possible to know exactly which are the true
  differentially abundant features (DAF).

  \item{true_DAF}{ a character vector. The names of the true DAF.}
  \item{FC_true_DAF}{ a numeric vector. The fold-change of the true DAF.}
}
\description{
Create a list that will contain all data needed for differential
features analysis.
}
\details{
If the columns names of the count table do not match the rows names of the
metadata table, metadata columns are screened for a variable countaining these names.

If such variable is not found, it is assumed that individuals follow the same order
through count table columns and metadata rows.
}
