% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{curve_plot}
\alias{curve_plot}
\title{Plot a ROC or PR curve}
\usage{
curve_plot(curve_data, type = c("roc", "pr"), plot_title = "[Title]",
  auc_limit = 1)
}
\arguments{
\item{res_data}{a data frame. The ouput of the \code{\link{generate_curve_data}} function}
}
\value{
a ggplot object with the ROC curve data
}
\description{
Plot a ROC or PR curve
}
