% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_curve_data.R
\name{generate_curve_data}
\alias{generate_curve_data}
\title{Compute metrics needed for ROC and PR curves}
\usage{
generate_curve_data(dt, threshold_var = "adjusted_pvalue")
}
\arguments{
\item{dt}{a data frame. The output of the \code{\link{run_DAF_analysis}} function.
If the function return a list countaining both "raw" and "curated" results,
then curated results must be used.}

\item{threshold_var}{a character. The name of the column that should
be used as threshold variable. By default "adjusted_pvalue".}
}
\value{
a list of 2 data frame :
\item{roc}{Countains TPR, FPR, method name and run number for each threshold value}
\item{pr}{Countains TPR, precision, method name and run number for each threshold value}
}
\description{
Based on results from differentially abundant features (DAF) analysis and a
variable used for threshold definition, it will compute 3 differents metrics,
allowing to plot ROC and precision-recall curves.
}
