% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_differences.R
\name{generate_differences}
\alias{generate_differences}
\title{Generate artificial dataset with differences between 2 groups}
\usage{
generate_differences(count_table, ...)
}
\arguments{
\item{count_table}{a matrix. Used as the \code{inputCount} paramter of the
\code{\link{EDDA::generateData}} function}

\item{...}{additional parameters for the \code{\link{EDDA::generateData}} function}
}
\value{
a list of  4 elements :
  \item{count_table}{ a matrix of count values.}
  \item{metadata}{ a data frame of at least one column countaining the group variable.}
  \item{true_DAF}{ a character vector. The names of the true DAF.}
  \item{FC_true_DAF}{ a numeric vector. The fold-change of the true DAF.}
}
\description{
Use the \code{\link{EDDA::generateData}} function to generate artificial count values table
}
