% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_repeated_DAF_data.R
\name{generate_repeated_DAF_data}
\alias{generate_repeated_DAF_data}
\title{Generate several dataset for DAF analysis}
\usage{
generate_repeated_DAF_data(count_table, metadata, times = 10, ...)
}
\arguments{
\item{count_table}{a matrix of count values with features as rows and samples as columns.}

\item{metadata}{a data frame. Rows are samples and miscellaneous variables as columns.}

\item{times}{Number of repetitions of the DAF analysis.}

\item{...}{additionnal parameter passed to the \code{\link{generateData}} function}
}
\value{
a list, each element being the output of a build_DAF_data function call
}
\description{
Call the \code{\link{build_DAF_data}} function several times and aggregate
all results ina list. Allow repeated and compared-repeated analysis
}
