% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMP_2012_metadata}
\alias{HMP_2012_metadata}
\title{HMP 2012 metadata}
\format{An object of class \code{data.frame} with 147 rows and 18 columns.}
\source{
\url{https://www.bioconductor.org/packages/3.4/data/experiment/html/curatedMetagenomicData.html}
}
\usage{
HMP_2012_metadata
}
\description{
Human Microbiome Project 2012 - Metadata provided with the
\code{curatedMetagenomicData} package.
}
\keyword{datasets}
