% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMP_2012_table_count}
\alias{HMP_2012_table_count}
\title{HMP 2012 table count}
\format{An object of class \code{matrix} with 1988 rows and 147 columns.}
\source{
\url{https://www.bioconductor.org/packages/3.4/data/experiment/html/curatedMetagenomicData.html}
}
\usage{
HMP_2012_table_count
}
\description{
Human Microbiome Project 2012 - Count table provided with the
\code{curatedMetagenomicData} package.
}
\keyword{datasets}
