% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_DAF_data.R
\name{preprocess_DAF_data}
\alias{preprocess_DAF_data}
\title{Filter DAF data}
\usage{
preprocess_DAF_data(data, taxon_lvl = c("g", "s", "t", "f", "o", "p",
  "c", NULL))
}
\arguments{
\item{data}{list. the output of the \code{\link{build_DAF_data}} function}
}
\value{
a list with the exact same format as the input data param
}
\description{
Remove some rows and columns from the count data table so that further process
is made easier
}
