% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ALDEx2.R
\name{process_ALDEx2}
\alias{process_ALDEx2}
\title{Process the DAF analysis through the ALDEx2 package}
\usage{
process_ALDEx2(data, ...)
}
\arguments{
\item{data}{the ouput of the \code{\link{build_DAF_data}} function}

\item{...}{additionnal parameters of the method}
}
\value{
a list countaining the raw output of ALDEx2 analysis and a curated version
}
\description{
Process the DAF analysis through the ALDEx2 package
}
