% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_edgeR.R
\name{process_edgeR}
\alias{process_edgeR}
\title{Process the DAF analysis through the edgeR package}
\usage{
process_edgeR(data, ...)
}
\arguments{
\item{data}{the ouput of the \code{\link{build_DAF_data}} function}

\item{...}{additionnal parameters of the method}
}
\value{
a list countaining the raw output of edgeR analysis and a curated version
}
\description{
Process the DAF analysis through the edgeR package
}
