% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mbzinb.R
\name{process_mbzinb}
\alias{process_mbzinb}
\title{Process the DAF analysis through the mbzinb package}
\usage{
process_mbzinb(data, ...)
}
\arguments{
\item{data}{the ouput of the \code{\link{build_DAF_data}} function}

\item{...}{additionnal parameters of the method}
}
\value{
a list countaining the raw output of mbzinb analysis and a curated version
}
\description{
Process the DAF analysis through the mbzinb package
}
