% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metagenomeSeq.R
\name{process_metagenomeSeq}
\alias{process_metagenomeSeq}
\title{Process the DAF analysis through the metagenomeSeq package}
\usage{
process_metagenomeSeq(data, ...)
}
\arguments{
\item{data}{the ouput of the \code{\link{build_DAF_data}} function}

\item{...}{additionnal parameters of the method}
}
\value{
a list countaining the raw output of metagenomeSeq analysis and a curated version
}
\description{
Process the DAF analysis through the metagenomeSeq package
}
