% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_RAIDA.R
\name{process_RAIDA}
\alias{process_RAIDA}
\title{Process the DAF analysis through the RAIDA package}
\usage{
process_RAIDA(data, ...)
}
\arguments{
\item{data}{the ouput of the \code{\link{build_DAF_data}} function}

\item{...}{additionnal parameters of the method}
}
\value{
the output of the DESeq() function processed through the results() function
}
\description{
Process the DAF analysis through the RAIDA package
}
