% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_voom.R
\name{process_voom}
\alias{process_voom}
\title{Process the DAF analysis through the vomm method of the limma package}
\usage{
process_voom(data, ...)
}
\arguments{
\item{data}{the ouput of the \code{\link{build_DAF_data}} function}

\item{...}{additionnal parameters of the method}
}
\value{
a list countaining the raw output of voom analysis and a curated version
}
\description{
Process the DAF analysis through the vomm method of the limma package
}
