% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_wilcox.R
\name{process_wilcox}
\alias{process_wilcox}
\title{Process the DAF analysis through a wilcoxon test}
\usage{
process_wilcox(data, ...)
}
\arguments{
\item{data}{the ouput of the \code{\link{build_DAF_data}} function}

\item{...}{additionnal parameters of the method}
}
\value{
the output of the wilcox test for each feature
}
\description{
Process the DAF analysis through a wilcoxon test
}
