% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ZIBseq.R
\name{process_ZIBseq}
\alias{process_ZIBseq}
\title{Process the DAF analysis through the ZIBseq package}
\usage{
process_ZIBseq(data, ...)
}
\arguments{
\item{data}{the ouput of the \code{\link{build_DAF_data}} function}

\item{...}{additionnal parameters of the method}
}
\value{
the rawouput and processed information of DAF
}
\description{
Process the DAF analysis through the ZIBseq package
}
