% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QinJ_2012_metadata}
\alias{QinJ_2012_metadata}
\title{Qin 2012 metadata}
\format{An object of class \code{data.frame} with 363 rows and 21 columns.}
\source{
\url{https://www.bioconductor.org/packages/3.4/data/experiment/html/curatedMetagenomicData.html}
}
\usage{
QinJ_2012_metadata
}
\description{
Metadata of the Qin2012 dataset provided with the
\code{curatedMetagenomicData} package.
}
\details{
Study of the link between microbiome and type 2 diabetes.
}
\keyword{datasets}
