% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QinJ_2012_table_count}
\alias{QinJ_2012_table_count}
\title{Qin 2012 table count}
\format{An object of class \code{matrix} with 1588 rows and 363 columns.}
\source{
\url{https://www.bioconductor.org/packages/3.4/data/experiment/html/curatedMetagenomicData.html}
}
\usage{
QinJ_2012_table_count
}
\description{
Count table of the Qin2012 dataset provided with the
\code{curatedMetagenomicData} package.
}
\details{
Study of the link between microbiome and type 2 diabetes.
}
\keyword{datasets}
