% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_features.R
\name{rank_features}
\alias{rank_features}
\title{Compare methods features rank}
\usage{
rank_features(results, adjusted = TRUE)
}
\arguments{
\item{results}{curated results from DAF analysis run}

\item{adjusted}{logical. should adjusted p-value be used ? default is TRUE}
}
\value{
a tibble of features names, method name and rank of the feature pvalue
}
\description{
Uses processed results object to order the features based on their p-value
}
