% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DAF_analysis.R
\name{run_DAF_analysis}
\alias{run_DAF_analysis}
\title{Run differentially abundant features analysis.}
\usage{
run_DAF_analysis(data, method = c("deseq2", "mbzinb", "zibseq", "raida",
  "aldex2", "metagenomeseq", "edger", "voom", "wilcox"), verbose = FALSE,
  ...)
}
\arguments{
\item{data}{a list. the returned object of the \code{\link{build_DAF_data}} function}

\item{method}{The name of the method to use for DAF detection. Any unambiguous
abbreviation of "deseq2", "mbzinb", "zibseq", "raida", "aldex2", "metagenomeseq",
"edger", "voom" or "wilcox". Default is "deseq2".}

\item{verbose}{logical. Should processing function display all of their messages}

\item{...}{additional parameters to be passed to the processing function.}
}
\value{
Data frame containing the pvalue and the adjusted pvalue (BH)
of the differential analysis as well as the names of methods and number of runs.
}
\description{
Call a specific processing function depending on the method chosen. Input data
are generated by the \code{\link{build_DAF_data}} function.
}
\details{
The function will first assess if the input data is a single dataset
or a list of several of them. Then it will check is one or more methods are called.
Based on these information, several kind of analysis are performed : a simple one,
a repeated one (focus on one method and average results), a compared one (multiple
methods called one time) or a compared and repeated one (several methods, several times).
}
