% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_process.R
\name{select_process}
\alias{select_process}
\title{Select processing function to call}
\usage{
select_process(method)
}
\arguments{
\item{method}{character. The name of the method for DAF}
}
\value{
a function performing the chosen analysis
}
\description{
Simple function to convert the method name (character) to the right function
}
