% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZellerG_2014_metadata}
\alias{ZellerG_2014_metadata}
\title{Zeller 2014 study metadata}
\format{An object of class \code{data.frame} with 199 rows and 22 columns.}
\source{
\url{https://www.bioconductor.org/packages/3.4/data/experiment/html/curatedMetagenomicData.html}
}
\usage{
ZellerG_2014_metadata
}
\description{
Metadata of the Zeller2014 dataset provided with the
\code{curatedMetagenomicData} package.
}
\details{
Study of the link between microbiome and colorectal cancer.
}
\keyword{datasets}
