% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZellerG_2014_table_count}
\alias{ZellerG_2014_table_count}
\title{Zeller 2014 count table}
\format{An object of class \code{matrix} with 1585 rows and 199 columns.}
\source{
\url{https://www.bioconductor.org/packages/3.4/data/experiment/html/curatedMetagenomicData.html}
}
\usage{
ZellerG_2014_table_count
}
\description{
Count table of the Zeller2014 dataset provided with the
\code{curatedMetagenomicData} package.
}
\details{
Study of the link between microbiome and colorectal cancer.
}
\keyword{datasets}
